Add super controller and model for objects
Family, user, artifact, and event are now concluded as "objects", and super controller and model are created for these 4. 
Simply specify the name of the database you want to manipulate as method argument, all of these 4 are accessible using the super controller methods.
Todo: deleting any record should also delete the relationships it has.
Passed tests of certain POST information.