Change the way of getting id_name
id_name used to be passed by a HTTP POST. It's now an argument of related methods.
Objects and Objects_model are treated as abstract superclass from now, only methods overriden in subclasses will be used by front end. This change is to reduce the information required for front end to pass.